+++ src/mainwindow.h
@@ -115,8 +115,8 @@ private:
 
     void updateGauge(const float &downValue, const float &upValue);
 
-    static QString convertSize(const int &size);
-    static QString convertSizeUnit(const int &size);
+    static QString convertSize(const uint64_t &size);
+    static QString convertSizeUnit(const uint64_t &size);
 
     // QWidget interface
 protected:
